Revert "moving sync to separate command"
This reverts commit ee973bbe93842d8b79e47ac95b6d70958402cdaa.
